Move cypress to /client
