+++ client/.eslintrc.json
@@ -3,7 +3,8 @@
   "extends": [
     "../.eslintrc.json",
     "plugin:react/recommended",
-    "plugin:react-hooks/recommended"
+    "plugin:react-hooks/recommended",
+    "plugin:cypress/recommended"
   ],
   "parser": "@typescript-eslint/parser",
   "env": {
