+++ client/cypress/fixtures/route-load-tests-config.js
@@ -1,4 +1,4 @@
-module.exports.route_load_tests_config = [
+export const route_load_tests_config = [
   {
     name: "Always failing route, to test error boundary",
     route: "error-boundary-test",
