+++ package.json
@@ -9,9 +9,7 @@
     "infobase_tmux_init": "npm run tmux_env",
     "tmux_env": "sh scripts/dev_scripts/tmux_env.sh",
     "refresh_pipeline_data": "sh scripts/dev_scripts/refresh_pipeline_data.sh",
-    "prepare": "husky install",
-    "cypress:open": "cypress open",
-    "cypress:run": "cypress run"
+    "prepare": "husky install"
   },
   "authors": [
     "Alex Cousineau-Leduc",
@@ -20,18 +18,16 @@
   ],
   "license": "MIT",
   "devDependencies": {
-    "axe-core": "^4.2.0",
     "babel-eslint": "^10.1.0",
-    "cypress": "^6.9.1",
-    "cypress-axe": "^0.12.2",
     "eslint": "^7.12.1",
     "eslint-plugin-cypress": "^2.11.2",
     "eslint-plugin-import": "^2.22.1",
     "eslint-plugin-lodash": "^7.1.0",
     "eslint-plugin-node": "^11.0.0",
     "husky": "^6.0.0",
-    "prettier": "^2.3.0",
     "pg": "^7.12.1",
-    "pg-hstore": "^2.3.3"
-  }
+    "pg-hstore": "^2.3.3",
+    "prettier": "^2.3.0"
+  },
+  "dependencies": {}
 }
